+++ src/main/java/mil/nga/proj/ProjectionFactoryType.java
@@ -19,6 +19,13 @@ public enum ProjectionFactoryType {
 	DEFINITION,
 
 	/**
+	 * Well-Known Text Definition parsing, PROJ conversion, and proj4j creation
+	 * 
+	 * @since 1.1.0
+	 */
+	DEFINITION_PARAMETERS,
+
+	/**
 	 * proj4j creation from a well-known coordinate reference system name
 	 */
 	NAME,