+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * Additional CRS Parser Swiss Oblique Mercator support
 * CRS Parser SCALE_FACTOR_ON_PSEUDO_STANDARD_PARALLEL support
 * CRS Parser CO_LATITUDE_OF_CONE_AXIS support
+* CRS Definition to PROJ parameters conversion and projection creation
 
 ## [1.0.0](https://github.com/ngageoint/projections-java/releases/tag/1.0.0) (06-24-2021)